Refactor: stat 조회 기능을 stat 모듈로 이동
