--- src/rank/rank.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { RankController } from './rank.controller';
-import { RankService } from './rank.service';
-
-@Module({
-    controllers: [RankController],
-    providers: [RankService],
-})
-export class RankModule {}
